TODOs have been completed with the needed files.

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,11 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // Due to potential accuracy and rounding challenges, floating-point numbers are not suitable for accurately representing monetary values like commission rates or discounts.
+            // Floats utilize binary representation for numbers, which can lead to inaccuracies during calculations involving decimal values.
+            // For accurate representation of monetary amounts, it is recommended to opt for decimal data types over floats.
+            // In the context of Laravel, you can employ the decimal method to define such columns in your migration file.
+            $table->decimal('commission_rate', 2, 2);
             $table->string('discount_code');
             $table->timestamps();
         });
