+++ database/migrations/2022_05_16_143445_create_orders_table.php
@@ -16,11 +16,13 @@ return new class extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
+            // $table->string('external_order_id');
             $table->foreignId('merchant_id')->constrained();
             $table->foreignId('affiliate_id')->nullable()->constrained();
             // TODO: Replace floats with the correct data types (very similar to affiliates table)
-            $table->float('subtotal');
-            $table->float('commission_owed')->default(0.00);
+            // Floats have been replaced with the decimal.
+            $table->decimal('subtotal', 2, 2);
+            $table->decimal('commission_owed', 2, 2)->default(0.00);
             $table->string('payout_status')->default(Order::STATUS_UNPAID);
             $table->string('discount_code')->nullable();
             $table->timestamps();
